+++ slidershow/menu.js
@@ -9,8 +9,8 @@ class Menu {
 
         this.$start = $("#start").focus().click(() => this.start_playback())
 
+        /** @type {Playback} */
         playback = this.playback = new Playback(this.aux_window) // expose global `playback`
-
 
         if (!$(FRAME_SELECTOR).length) {
             this.$start_wrapper.hide()
@@ -123,16 +123,18 @@ class Menu {
     }
 
     display_progress(max, $placement = null) {
-        $("#progress").remove()
         const $progress = $("<div/>", { id: "progress" }).insertAfter($placement || "h1").circleProgress({
             value: 0,
             max: max
         })
         let progress = 0
-        return () => {
+        return (finish=false) => {
+            if(finish) {
+                progress = max-1
+            }
             $progress.circleProgress("value", ++progress)
             if (progress === max) {
-                $("#progress").fadeOut(2000)
+                $progress.fadeOut(2000, ()=>$progress.remove())
             }
         }
     }
@@ -177,25 +179,39 @@ class Menu {
     }
 
     async export(compact_file = false, path = "") {
-        const $contents = $($main.prop('outerHTML'))
+        // Why to wrap the body inside a div? jQuery seems to handle such fundamental tags differently.
+        // We end up with a collection of body children, not with the body itself.
+        const $contents = $("<div>" + $("body").prop('outerHTML') + "</div>")
 
         // reduce parameters
         $contents.removeAttr("style")
         $contents.find("*").removeAttr("style")
+        $contents.find("> #map, > #map-hud, > #map-wrapper, > #hud, > menu, > .ZebraDialog, > .ZebraDialogBackdrop").remove()
         await Frame.finalize_frames($contents, this.playback.$articles, compact_file, path, this.display_progress(this.playback.$articles.length))
 
-        const html = $contents.prop("outerHTML")
+        const html = $contents.prop("innerHTML")
         if (!html.length) {
             this.playback.hud.alert("Cannot export a single file – too big.")
         }
-        const data = `<html>\n<head><meta charset="UTF-8">
-<script src="${DIR}slidershow.js"></script>
-</head>\n<body>` + html + "\n</body>\n</html>"
+
+        // Prepare the original head.
+        // Why not use HEAD=document.head.innerHTML cached at the application start in slidershow.js?
+        // As the document is not ready yet, it would not show whole header but only the part the parser is it.
+        // Consider this head: `<script slidershow.js><link href=custom.css>`
+        // While accessing document.head in slidershow.js, link is still invisible. The slidershow.js needed to be
+        // the very last tag in the head. Which would cover different problems:
+        // CSS order, user might want to add another script accessing slidershow properties...
+        let $head = $("<div>" + $("head").prop('outerHTML')+ "</div>")
+        $head.find("[data-templated]").remove() // remove all dynamically added libraries
+        $head.find("[src^='https://api.mapy.cz'],[href^='https://api.mapy.cz']").remove() // including vendor libraries that does not our honour [data-templated] attr
+
+        // Export the data blob
+        const data = `<!DOCTYPE html><html><head>\n${$head[0].innerHTML}</head>\n<body>` + html + "\n</body>\n</html>"
         const blob = new Blob([data.replaceAll(EXPORT_SRC, "src")], { type: "text/plain" })
         const url = URL.createObjectURL(blob)
         const link = document.createElement("a")
         link.href = url
-        link.download = "slidershow.html"
+        link.download = this.playback.session.docname
         document.body.appendChild(link)
         link.click()
         URL.revokeObjectURL(url)