+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * auxiliary window to track the presentation pace
 * thumbnail ribbon
 * support for various media formats
+* header and footer template
+* stable save
+* F5 honours the current frame (hash)
 
 ## 0.7.0
 Every important feature working as expected.